+++ src/main/java/nextstep/subway/domain/Station.java
@@ -1,11 +1,14 @@
 package nextstep.subway.domain;
 
+import lombok.Getter;
+
 import javax.persistence.Entity;
 import javax.persistence.GeneratedValue;
 import javax.persistence.GenerationType;
 import javax.persistence.Id;
 
 @Entity
+@Getter
 public class Station {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
@@ -19,11 +22,8 @@ public class Station {
         this.name = name;
     }
 
-    public Long getId() {
-        return id;
-    }
-
-    public String getName() {
-        return name;
+    public Station(Long id, String name) {
+        this.id = id;
+        this.name = name;
     }
 }
