+++ src/test/java/nextstep/subway/acceptance/line/LineAcceptanceTest.java
@@ -11,9 +11,9 @@ import org.springframework.http.MediaType;
 import java.util.HashMap;
 import java.util.Map;
 
-import static nextstep.subway.acceptance.line.LineSteps.지하철_노선_목록_조회_요청;
-import static nextstep.subway.acceptance.line.LineSteps.지하철_노선_생성_요청;
-import static nextstep.subway.acceptance.line.LineSteps.지하철_노선_조회_요청;
+import static nextstep.subway.acceptance.step.LineSteps.지하철_노선_목록_조회_요청;
+import static nextstep.subway.acceptance.step.LineSteps.지하철_노선_생성_요청;
+import static nextstep.subway.acceptance.step.LineSteps.지하철_노선_조회_요청;
 import static nextstep.subway.utils.RestAssuredStep.given;
 import static org.assertj.core.api.Assertions.assertThat;
