+++ src/test/java/nextstep/subway/acceptance/member/MemberAcceptanceTest.java
@@ -10,7 +10,7 @@ import org.springframework.http.HttpStatus;
 
 import static nextstep.subway.utils.AdministratorInfo.ADMIN_EMAIL;
 import static nextstep.subway.utils.AdministratorInfo.ADMIN_PASSWORD;
-import static nextstep.subway.acceptance.member.MemberSteps.*;
+import static nextstep.subway.acceptance.step.MemberSteps.*;
 import static org.assertj.core.api.Assertions.assertThat;
 
 class MemberAcceptanceTest extends AcceptanceTest {
