+++ src/test/java/nextstep/subway/acceptance/section/SectionAcceptanceTest.java
@@ -11,8 +11,8 @@ import org.springframework.http.HttpStatus;
 import java.util.HashMap;
 import java.util.Map;
 
-import static nextstep.subway.acceptance.line.LineSteps.*;
-import static nextstep.subway.acceptance.station.StationSteps.지하철역_생성_요청;
+import static nextstep.subway.acceptance.step.LineSteps.*;
+import static nextstep.subway.acceptance.step.StationSteps.지하철역_생성_요청;
 import static org.assertj.core.api.Assertions.assertThat;
 
 @DisplayName("지하철 구간 관리 기능")
