+++ src/test/java/nextstep/subway/acceptance/step/MemberSteps.java
@@ -1,4 +1,4 @@
-package nextstep.subway.acceptance.member;
+package nextstep.subway.acceptance.step;
 
 import io.restassured.RestAssured;
 import io.restassured.response.ExtractableResponse;
